+++ node/src/components/consensus/protocols.rs
@@ -1,3 +1,3 @@
 pub(crate) mod common;
 pub(crate) mod highway;
-pub(crate) mod simple_consensus;
+pub(crate) mod zug;